advanced/intelligent_customer_assistant/requirements-improved.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
pydantic==2.4.2
pyyaml==6.0.1
gunicorn==21.2.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pymysql==1.1.0